# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Keyword alert API",
    "description": "sample API.",
    "termsOfService": "https://github.com/jhwmon/express-keyword-alert",
    "contact": {
      "email": "jmonaco88@gmail.com"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
...
  "paths": {
...
    "/comments": {
      "get": {
        "tags": ["Comments"],
        "summary": "댓글 목록",
        "operationId": "findComments",
        "parameters": [
          {
            "name": "boardId",
            "in": "query",
            "description": "게시글 번호",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "페이지 번호(0부터 시작함)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "페이지당 사이즈",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "total": {
                          "type": "number"
                        },
                        "totalPages": {
                          "type": "number"
                        },
                        "currentPage": {
                          "type": "number"
                        },
                        "comments": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer"
                              },
                              "boardId": {
                                "type": "integer"
                              },
                              "msg": {
                                "type": "string"
                              },
                              "owner": {
                                "type": "string"
                              },
                              "createdAt": {
                                "type": "string"
                              },
                              "updatedAt": {
                                "type": "string"
                              },
                              "subComments": {
                                "type": "object",
                                "properties": {
                                  "id": {
                                    "type": "integer"
                                  },
                                  "pcId": {
                                    "type": "integer"
                                  },
                                  "msg": {
                                    "type": "string"
                                  },
                                  "owner": {
                                    "type": "string"
                                  },
                                  "createdAt": {
                                    "type": "string"
                                  },
                                  "updatedAt": {
                                    "type": "string"
                                  }
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {}
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-codegen-request-body-name": "body"
      },
      "post": {
        "tags": ["Comments"],
        "summary": "댓글 추가",
        "operationId": "createComments",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "contents", "owner", "pwd"],
                "properties": {
                  "boardId": {
                    "type": "integer"
                  },
                  "msg": {
                    "type": "string"
                  },
                  "owner": {
                    "type": "string"
                  },
                  "pcId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "댓글 추가 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "댓글 추가 성공"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "comments": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "example": 1
                            },
                            "msg": {
                              "type": "string"
                            },
                            "owner": {
                              "type": "string",
                              "example": "jhwmon"
                            },
                            "pcId": {
                              "type": "integer"
                            },
                            "createdAt": {
                              "type": "string",
                              "example": "2022-04-24 12:00:00"
                            },
                            "updatedAt": {
                              "type": "string",
                              "example": "2022-04-24 13:01:02"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "잘못 된 요청"
                    }
                  }
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  }
}